Restructure login logic and add login state persistence
