+++ backend/db/queries/users.js
@@ -1,7 +1,7 @@
 const db = require('../connection');
 
 const getLogin = (email, password) => {
-  const queryString = `SELECT email, password FROM users
+  const queryString = `SELECT email, password, * FROM users
   WHERE email = $1
   AND password = $2;
   `;
