+++ frontend/src/App.js
@@ -1,12 +1,12 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import './App.css';
 
 //Components
 import LoginPage from './components/LoginPage';
 import Signup from './components/Signup';
 import NavBar from './components/NavBar';
-
+import Profile from './components/Profile';
 
 function App() {
   const [login, setLogin] = useState(false);
@@ -14,69 +14,26 @@ function App() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
 
-  //Handle login state
-  const loggedIn = async (e) => {
-    if (e === null) {
-      setLogin(false);
-      return;
-    }
-
-    e.preventDefault();
-    console.log(email, password);
-
-    try {
-      const body = { email, password };
-      console.log(body);
-      const response = await fetch('/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body)
-      });
-
-      if (!response.ok) {
-        const errorData = await response.json();
-        throw new Error(errorData.error);
-      }
-
-      const data = await response.json();
-      console.log(data);
+  useEffect(() => {
+    const user = window.localStorage.getItem('userid');
+    if (user !== null) {
       setLogin(true);
-      setEmail('');
-      setPassword('');
-      setError('');
-    } catch (err) {
-      console.error(err.message);
-      setError(err.message);
-      setEmail('');
-      setPassword('');
     }
-  };
-
+  }, []);
+  
 
   return (
     <BrowserRouter>
       <div className="App">
 
-        <NavBar login={login} loggedIn={loggedIn} />
+        <NavBar login={login} setLogin={setLogin} />        
 
         <Routes>
-          {
-            login ? <></> :
-              <Route path="/"
-                element={
-                  <LoginPage
-                    loggedIn={loggedIn}
-                    email={email}
-                    setEmail={setEmail}
-                    password={password}
-                    setPassword={setPassword}
-                    error={error}
-                  />}
-              />
-          }
+          <Route path='/profile' element={<Profile />} />
+          <Route path="/" element={<LoginPage email={email} setEmail={setEmail} password={password} setPassword={setPassword} error={error} setError={setError} setLogin={setLogin} />} />
           <Route path="/signup" element={<Signup setLogin={setLogin} />} />
-
         </Routes>
+
       </div>
     </BrowserRouter>
   );
